add the generic settings to the component folder
