+++ app/src/SettingsTabs/PointSettingsPopup.tsx
@@ -1,11 +1,11 @@
 import { Box, Button, FormControlLabel, Modal, TextField, TextFieldProps, Typography } from "@mui/material";
-import { ColumnInfo, PointSettings, assertPointSettings } from "@tsconline/shared";
+import { ColumnInfo, assertPointSettings } from "@tsconline/shared";
 import { observer } from "mobx-react-lite";
 import { useContext, useState } from "react";
 import { context } from "../state";
-import { NumericFormat } from "react-number-format";
 import "./PointSettingsPopup.css";
 import { TSCButton, TSCCheckbox } from "../components";
+import { GenericTextField } from "../components";
 
 type PointSettingsPopupProps = {
   column: ColumnInfo;
@@ -96,41 +96,5 @@ export const PointSettingsPopup: React.FC<PointSettingsPopupProps> = observer(({
         </Box>
       </Modal>
     </div>
-  );
-});
-
-type GenericTextFieldProps = {
-  header: string;
-  inputs: {
-    helperText: string;
-    value: number;
-    onValueChange: (value: number) => void;
-  }[];
-};
-
-const GenericTextField: React.FC<GenericTextFieldProps> = observer(({ header, inputs }) => {
-  const InputTextField = ({ ...props }: TextFieldProps) => <TextField {...props} className="generic-text-field" />;
-
-  return (
-    <Box className="generic-text-field-container">
-      <Typography className="generic-text-field-header">{header}</Typography>
-      <div className="generic-text-fields">
-        {inputs.map((input, index) => (
-          <NumericFormat
-            key={index}
-            helperText={input.helperText}
-            value={input.value}
-            customInput={InputTextField}
-            onValueChange={(values) => {
-              const floatValue = values.floatValue;
-              if (!floatValue) {
-                return;
-              }
-              input.onValueChange(floatValue);
-            }}
-          />
-        ))}
-      </div>
-    </Box>
   );
 });
