+++ app/src/components/TSCGenericSettings.css
@@ -1,0 +1,27 @@
+
+.generic-text-field-header {
+  display: flex;
+  justify-content: center;
+  text-align: center;
+  margin-bottom: 5px;
+  font-weight: 600;
+}
+
+.generic-text-field-container {
+  display: flex;
+  flex-direction: column;
+  gap: 10px;
+  text-align: center;
+}
+
+.generic-text-field {
+  padding-left: 10px;
+  padding-right: 10px;
+  width: 123px;
+}
+
+.generic-text-fields {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
